Upgraded to VS2022

+++ gpftw/ftw_array_view.h
@@ -23,7 +23,7 @@ namespace ftwl
 		template <size_t _elementCount>
 		inline constexpr							array_view					(_tElement (&_dataElements)[_elementCount])									noexcept	: Data(_dataElements), Count(_elementCount)										{}
 		template <size_t _elementCount>
-		inline constexpr							array_view					(_tElement (&_dataElements)[_elementCount], uint32_t elementCount)			noexcept	: Data(_dataElements), Count(::ftwl::min(_elementCount, elementCount))			{}
+		inline constexpr							array_view					(_tElement (&_dataElements)[_elementCount], uint32_t elementCount)			noexcept	: Data(_dataElements), Count(::ftwl::min((uint32_t)_elementCount, elementCount))			{}
 
 		// Operators
 							_tElement&				operator[]					(uint32_t index)																		{ if(0 == Data) throw(::std::exception("Uninitialized array pointer.")); if(index >= Count) throw(::std::exception("Invalid index.")); return Data[index]; }
